Changing the name on the release tag since the first LSRD release will be version 1.0.0 and we still want to tag the code we received from Feng.

+++ ledapsSrc/src/lndcal/input.c
@@ -1076,7 +1076,7 @@ bool GetHeaderInput(Input_t *this, char *file_header_name, Param_t *param) {
           this->meta.gain_set[i] = (Gain_t)
 	    KeyString(key.value[i], key.len_value[i],
 	              Gain_string, (int)GAIN_NULL, (int)GAIN_MAX);
-          if (this->meta.inst == (int)GAIN_NULL) {
+          if (this->meta.inst == GAIN_NULL) {
             error_string = "invalid gain setting string";
 	    break;
 	  }
